requirements.txt
anthropic>=0.39.0
requests>=2.32.0
feedparser>=6.0.11
pyyaml>=6.0.2
python-dotenv>=1.0.1
